Flatten repository folder
