+++ mcc_generated_files/device_config.c
