+++ mcc_generated_files/pin_manager.h
