+++ nbproject/Makefile-genesis.properties
@@ -1,0 +1,10 @@
+#
+#Sun Jul 19 14:58:54 MDT 2020
+default.Pack.dfplocation=C\:\\Program Files (x86)\\Microchip\\MPLABX\\v5.30\\packs\\Microchip\\PIC18F-K_DFP\\1.2.72
+default.languagetoolchain.dir=C\:\\Program Files (x86)\\Microchip\\xc8\\v2.05\\bin
+configurations-xml=b18ce8a0fe6e21e1ac7e8856f143e99e
+com-microchip-mplab-nbide-embedded-makeproject-MakeProject.md5=7cd2eead2ea6964989cbf02efe721a76
+default.languagetoolchain.version=2.05
+host.platform=windows
+conf.ids=default
+default.com-microchip-mplab-nbide-toolchainXC8-XC8LanguageToolchain.md5=f87389d82b4b299d292110931953ebdc
