+++ nbproject/private/SuppressibleMessageMemo.properties
@@ -1,0 +1,4 @@
+#
+#Sun Jun 21 17:35:51 MDT 2020
+pk3/CHECK_4_HIGH_VOLTAGE_VPP=true
+pk3/DEVID_MISMATCH=true
